+++ tests/integration_tests/functional/test_metrics.py
@@ -2,8 +2,9 @@
 # SPDX-License-Identifier: Apache-2.0
 """Tests the metrics system."""
 
+import datetime
 import os
-import json
+import math
 import platform
 import host_tools.logging as log_tools
 
@@ -25,8 +26,7 @@ def test_flush_metrics(test_microvm_with_api):
 
     microvm.start()
 
-    res = metrics_fifo.sequential_reader(1)
-    metrics = json.loads(res[0])
+    metrics = microvm.flush_metrics(metrics_fifo)
 
     exp_keys = [
         'utc_timestamp_ms',
@@ -54,4 +54,11 @@ def test_flush_metrics(test_microvm_with_api):
 
     assert set(metrics.keys()) == set(exp_keys)
 
-    microvm.flush_metrics(metrics_fifo)
+    utc_time = datetime.datetime.now(datetime.timezone.utc)
+    utc_timestamp_ms = math.floor(utc_time.timestamp() * 1000)
+
+    # Assert that the absolute difference is less than 1 second, to check that
+    # the reported utc_timestamp_ms is actually a UTC timestamp from the Unix
+    # Epoch.Regression test for:
+    # https://github.com/firecracker-microvm/firecracker/issues/2639
+    assert abs(utc_timestamp_ms - metrics['utc_timestamp_ms']) < 1000